requirements.txt
pytesseract==0.2.4
pandas==0.23.3
Pillow==5.2.0
